COH-17463 RFA: DatagramTest has int rollover issues if run for several hours COH-17463 RFA: DatagramTest has int rollover issues if run for several hours (auto-submit 108824 after successfully running remote remote.full) Job ID: job.9.20240514195825.17001 (auto-submit integ 109002 main -> coherence-ce/release/coherence-ce-v24.03 after successfully running remote remote.full) Job ID: job.9.20240515142739.12510
[git-p4: depot-paths = "//dev/coherence-ce/release/coherence-ce-v24.03/": change = 109012]

+++ prj/coherence-core/src/main/java/com/tangosol/net/DatagramTest.java
@@ -1,8 +1,8 @@
 /*
- * Copyright (c) 2000, 2020, Oracle and/or its affiliates.
+ * Copyright (c) 2000, 2024, Oracle and/or its affiliates.
  *
  * Licensed under the Universal Permissive License v 1.0 as shown at
- * http://oss.oracle.com/licenses/upl.
+ * https://oss.oracle.com/licenses/upl.
  */
 
 package com.tangosol.net;
@@ -1530,8 +1530,8 @@ public class DatagramTest
                     long ldtSentNanos = stream.readLong();
                     long ldtAckNanos  = stream.readLong();
 
-                    int nCurrent = stream.readInt();
-                    int nBytes   = cbPayload < 0 ? packet.getLength() : Math.min(cbPayload, packet.getLength());
+                    long nCurrent = stream.readLong();
+                    int  nBytes   = cbPayload < 0 ? packet.getLength() : Math.min(cbPayload, packet.getLength());
 
                     lifeTracker.trackArrival(nCurrent, ldtSentNanos, ldtAckNanos, nBytes);
                     nowTracker.trackArrival(nCurrent, ldtSentNanos, ldtAckNanos, nBytes);
@@ -1541,7 +1541,7 @@ public class DatagramTest
                     for (int i = 7/*header length in ints*/, c = nProcess / 4; i < c; ++i)
                         {
                         // check for corrupted packet
-                        int n = stream.readInt();
+                        long n = stream.readLong();
                         if (n != nCurrent)
                             {
                             if (n == 0)
@@ -1664,7 +1664,7 @@ public class DatagramTest
         * @param nCurrent packet number
         * @param cBytes   packet size in bytes
         */
-        public void trackArrival(int nCurrent, long ldtSentNanos, long ldtAckNanos, int cBytes)
+        public void trackArrival(long nCurrent, long ldtSentNanos, long ldtAckNanos, int cBytes)
             {
             long ldtNow = System.currentTimeMillis();
